refactor(racingcar): model package를 판단 class와 변형 class로 나누기, 미구현 부분 제거
