+++ src/main/java/controller/RacingCarGameManager.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import view.InputView;
 import view.OutputView;
+import model.TransformRacingValue;
 import model.ValidateRacingValue;
 import utility.ValidateException;
 
@@ -13,14 +14,14 @@ public class RacingCarGameManager {
         List<Integer> raceResult = new ArrayList<>();
         String carNameBundle;
         int numberOfMatch;
+
         OutputView.racingCarName();
         carNameBundle = InputView.racingCarName();
         OutputView.numberOfRacing();
-        ValidateException.numberOfRacingInvalid(InputView.numberOfMatch());
         numberOfMatch = Integer.parseInt(InputView.numberOfMatch());
-        carName = ValidateRacingValue.carNameInputToList(carNameBundle);
+        carName = TransformRacingValue.carNameInputToList(carNameBundle);
         ValidateException.racingCarNameInvalid(carName);
-        raceResult = ValidateRacingValue.raceResultListReset(carName.size(), raceResult);
+        raceResult = TransformRacingValue.raceResultListReset(carName.size(), raceResult);
         for (int i = 0; i < numberOfMatch; i++) {
             ValidateRacingValue.racingResult(carName, raceResult);
             OutputView.racingResult(carName, raceResult);
