--- src/main/java/utility/ManageConstantValue.java
@@ -1,10 +1,0 @@
-package utility;
-
-public class ManageConstantValue {
-    public enum constantOutPutView {
-
-    }
-    public enum replaceMagicNumberToConstant {
-
-    }
-}
