+++ src/test/java/model/ValidateRacingValueTest.java
@@ -1,6 +1,6 @@
 package model;
 
-import static model.ValidateRacingValue.carNameInputToList;
+import static model.TransformRacingValue.carNameInputToList;
 import static model.ValidateRacingValue.racingResult;
 import static model.ValidateRacingValue.racingWinner;
 import static model.ValidateRacingValue.randomMoreFour;
@@ -13,16 +13,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class ValidateRacingValueTest {
-    @DisplayName("쉼표로 구분된 string이 List에 저장되는가")
-    @Test
-    void carNameInputToListTest() {
-        //given
-        String carName = "pobi,nana,popo";
-        //when
-        List<String> listCarName = new ArrayList<>(carNameInputToList(carName));
-        //then
-        assertThat(listCarName).contains("pobi", "nana", "popo");
-    }
+
     @DisplayName("숫자 랜덤 추출 기능이 작동하는가")
     @Test
     void randomMoreFourTest() {
@@ -32,6 +23,7 @@ class ValidateRacingValueTest {
         //then
         System.out.printf("랜덤 추출 기능 출력값 : %b \n", testResult);
     }
+
     @DisplayName("추가된 레이싱 결과값이 List에 저장되는가")
     @Test
     void racingResultTest() {
@@ -42,11 +34,12 @@ class ValidateRacingValueTest {
         List<Integer> testList = racingResult(carName, raceResult);
         //then
         System.out.print("레이싱 결과 저장 기능 출력값 :");
-        for(int result : testList) {
+        for (int result : testList) {
             System.out.printf("%d,", result);
         }
         System.out.println();
     }
+
     @DisplayName("결과 값에 따라 우승자를 선정하는가")
     @Test
     void racingWinnerTest() {
